Rewrite to make better use of make.
Instead of having the test runner do the extraction, make can do it for
us. This allows for better parallelization. It also lets the WADs be
extracted separately (and potentially modified) before the tests are
run; I'm planning to reuse this for wadptr so that I can test WADs play
back as intended after being compressed.